requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
mistralai>=1.4.0
googlemaps>=4.10.0
requests>=2.20.0
aiohttp>=3.8.0
pydantic>=2.9.0
typing-extensions>=4.12.2
httpx>=0.27.0
jsonpath-python>=1.0.6
python-dateutil>=2.8.2
typing-inspect>=0.9.0